Update code for windows platform

+++ src/lib/src/stream/AudioSender.cpp
@@ -22,7 +22,7 @@ int AudioSender::begin() {
 	}
 
 	Config::setInt(AUDIO_CHANNELS, m_ifmt_ctx->streams[0]->codecpar->ch_layout.nb_channels);
-	Config::setInt(AUDIO_BIT_RATE, m_ifmt_ctx->streams[0]->codecpar->bit_rate);
+	Config::setInt(AUDIO_BIT_RATE, (int)m_ifmt_ctx->streams[0]->codecpar->bit_rate);
 	Config::setInt(AUDIO_SAMPLE, m_ifmt_ctx->streams[0]->codecpar->sample_rate);
 	Config::setInt(AUDIO_CODEC, m_ifmt_ctx->streams[0]->codecpar->codec_id);
 	Config::setInt(AUDIO_FORMAT, m_codec->sample_fmts[0]);
